pyproject.toml
[tool.poetry]
name = "escapetown"
version = "0.1.0"
description = ""
authors = ["Tina Koziol <tinakoziol@yahoo.de>"]

[virtualenvs]
create = true
in-project = true

[tool.poetry.dependencies]
python = "^3.12"
Django = "4.2.7"
arrow = "1.3.0"
asgiref = "3.6.0"
binaryornot = "0.4.4"
certifi = "2023.7.22"
chardet = "5.2.0"
charset-normalizer = "3.3.1"
click = "8.1.7"
cookiecutter = "2.4.0"
django-browser-reload = "1.12.0"
django-tailwind = "3.6.0"
idna = "3.4"
jinja2 = "3.1.2"
markdown-it-py = "3.0.0"
markupsafe = "2.1.3"
mdurl = "0.1.2"
pygments = "2.16.1"
python-dateutil = "2.8.2"
python-slugify = "8.0.1"
pyyaml = "6.0.1"
requests = "2.31.0"
rich = "13.6.0"
six = "1.16.0"
sqlparse = "0.4.4"
text-unidecode = "1.3"
types-python-dateutil = "2.8.19.14"
urllib3 = "2.0.7"
pillow = "^10.3.0"
psycopg2-binary = "^2.9.9"
django-markdownx = "^4.0.7"
markdown = "^3.6"
django-markdown-deux = "^1.0.6"
markup = "^0.2.2"
reportlab = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
